requirements.txt
asgiref==3.4.1
autopep8==1.6.0
Django==3.2.12
djangorestframework==3.13.1
pkg-resources==0.0.0
pycodestyle==2.8.0
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
typing-extensions==4.1.1
